// src/payload-types.ts
export interface Message {
  id: number;
  name: string;
  email: string;
  message: string;
  ip: string;
  read?: boolean | null;
  type?: ('reply' | 'initial') | null;
  updatedAt: string;
  createdAt: string;
}